requirements.txt
functions-framework==3.*
pandas==2.*
numpy

google-auth==2.*
gspread==6.*
google-cloud-secret-manager==2.*
google-cloud-bigquery==3.*
google-cloud-bigquery-storage==2.*
db-dtypes
google-api-python-client==2.*
google-auth-httplib2
httplib2                               